Separate the functions in to two pages
