+++ requirements.txt
@@ -19,6 +19,7 @@ Deprecated==1.2.14
 dill==0.3.8
 dirtyjson==1.0.8
 distro==1.9.0
+docx2txt
 flatbuffers==24.3.7
 frozenlist==1.4.1
 fsspec==2024.2.0
